pyproject.toml
[tool.poetry]
name = "equine-microbiome-reporter"
version = "0.1.0"
description = "Automated PDF report generation for equine gut microbiome analysis from 16S rRNA sequencing data"
authors = ["HippoVet+"]
readme = "README.md"
packages = [{include = "*.py"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.13"
pandas = "^1.3.0"
matplotlib = "^3.4.0"
numpy = "^1.21.0"
reportlab = "^3.6.0"
PyYAML = "^6.0"
Flask = "^2.3.0"
Werkzeug = "^2.3.0"
jinja2 = "^3.1.6"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
